pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = "Shop"
authors = ["sudodebug <sudodebug@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
pillow = "^9.5.0"
celery = "5.2.7"
flower = "^2.0.0"
stripe = "^5.4.0"
weasyprint = "^59.0"
webencodings = "^0.5.1"
setuptools = "^68.0.0"
redis = "^4.6.0"
django-rosetta = "^0.9.9"
django-parler = "^2.3"
django-localflavor = "^4.0"
flake8 = "^6.1.0"
black = "^23.9.1"


[tool.poetry.group.dev.dependencies]
setuptools = "^68.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
